Attempted to make a "/shrug" like command.
But then I realised the bot can't edit other people's messages and I just settled for a regular "say this and the bot will just post the emoji" command.

Also, the "ask the gatekeeper to let you in" command now deletes the message with the command so as to not potentially clog whichever chat it's used in.

A bunny:
(\(\
( - -)
((') (')